优化 api points 访问地址

+++ module/module-sys/src/main/java/cn/yoyo/module/sys/domain/entity/SysConfig.java
@@ -20,7 +20,7 @@ public class SysConfig{
     /**
      * 参数主键
      */
-    private Integer configId;
+    private Long configId;
     /**
      * 参数名称
      */
